Use `float` for `updated_at` timestamp

+++ edb/server/protocol/auth_ext/base.py
@@ -42,7 +42,7 @@ class BaseProvider:
     async def fetch_emails(self, token: str) -> list[data.Email]:
         raise NotImplementedError
 
-    def _maybe_isoformat_to_timestamp(self, value: str | None) -> int | None:
+    def _maybe_isoformat_to_timestamp(self, value: str | None) -> float | None:
         return (
-            int(datetime.fromisoformat(value).timestamp()) if value else None
+            datetime.fromisoformat(value).timestamp() if value else None
         )
